commander: move battery handling into arming checks
